Work around PR status db object missing statuses property
As I spun up desktop@master on my Windows machine today I got this error

```
Uncaught (in promise) TypeError: Cannot read property 'map' of undefined
    at PullRequestStore.getPRStatusById (pull-request-store.ts:237)
    at <anonymous>
```

It seems the `statuses` property was added in #3588 and typed as `ReadonlyArray<IAPIRefStatusItem>` but since no corresponding database upgrade function was added to ensure that the property exists on prior database objects we can't rely on it being there.

I could add an upgrade function but seeing as we're pretty close to a release I would rather just mark the property accurately and make sure we don't have any other places which rely on it always being present.

We can revisit and clean this up in a later version.