Make menu bars themeable
The current setup for the menu bars on Windows/Linux was using the exact same styles as for the toolbar dropdowns which works great in the light theme and not so much in the dark theme.

This commits puts in place variables which by default inherit the toolbar colors and is customizeable in other themes.