Clone-Repository now verifies path on focus
Previously if one selected a folder path that already existed, an error would show up and remain even if a user then removed the folder.

Now when the user returns to the clone window and focus is captured it will check the path.

If the path exists, it will keep or throw the error and block cloning.

If the path no longer exists, it will clear the error and allow cloning.