Branch Exists On Remote warning now checks BranchType
Previously I had made a mistake in checking if the branch.remote variable was null to determine if the branch existed on remote.

This was effective for local branches that had then been published to  remote. But it was not successful for remote branches that were not also local branches.

Now it checks the BranchType to ensure the branch is a remote branch.

It also checks nameWithoutRemote so that it can catch both local branches that were published to remote as well as remote branches that show as 'remote/branchName'