Passed getNamesOfSelectedOversizedFiles() down to commit-message
Passed the function that retrieves the paths of files to be included in a commit that are over 100 MB down to Commit-Message via props.

This will allow calling this function when attempting to commit, so that a check can be done for any potential large file issues when Git LFS is not set up.