Retrieve necessary variables early to avoid duplication of effort
In both the cirumstance that the OversizedFilesWarning popup dialog is shown and the createCommit function immediately returns as well as simply continuing with the commit requires the same variables.

Previously each route retrieved them separately which was not necessary. Instead they are now retrieved earlier and are available for both situations.