Replaced uses of path.parse with path.basename
Use of path.parse led to an extra step of using .base afterwards to get the base name of the path.

This is not necessary as there is a specific function for this called basename which takes care of both.

Additionally this will match better with the rest of the codebase.