Moved checks for large files on committing into Sidebar.tsx
Previously this logic occurred in commit-message. The commit button is pressed by the user triggering a commit, and this logic occurred in that function.

However, commit-message does not have access to the working directory files, so additional functionality was added to the changes-list and passed down to commit-message.

Instead, the function that checks for large files has been moved into a lib file that can be exported, and will be accessed from the sidebar where this logic now is.

This allows avoiding the need to pass down functions via props, and allows for the file size check to be accessed in other places of the code base too.

Co-Authored-By: shiftkey <github@brendanforster.com>