readMeExists is now updated on window focus
Previously the state variable readMeExists did not update on window focus. This means that if a user switches from GitHub Desktop and adds or removes a ReadMe.md file to the folder, the warning would not update.

Now the readme will be detected on window focus so that it can display or remove the warning if the the file's existence has changed.