Simplified doesReadMeExist function
This function no longer uses fse.readdir.
Instead it makes use of the fse.pathExists function that always returns a boolean whether or not a specific file exists.

Previously a regular expression was used to detect files of any capitlization as long as the characters matched.

However, this is not necessary. On Linux it is possible to have several files in one directory with the same name but different capitalization. For Linux we only want to detect when the file is exactly "README.md"

However, for Windows and Mac OS, only one file can exist with a given name in a directory. Additionally, case does not matter. If a file has the same characters but different capitalization, it will be detected. This means we can find the file no matter what the captalization is by searching for a name with the right letters.