doesReadMeExist is now updateReadMeExists, sets state directly
Previously this function accepted two string arguments and returned a boolean value depending on if the given path contained a README.md file.

However, since the result is used to update the state of readMeExists, the function has been updated to control the state directly.

It still takes two arguments, since the path and name each may be overriding the state in their particular case. So we want to still give that control.