Commit summary clearing now waits until commit creation is complete
Previously the completion of the commit creation was not awaited. This means the popup would close and the commit summary would be cleared while it is committing.

The result of this was that the commit summary would now display as the default commit summary while it's committing, appearing as though it is not the entered summary.

This was done so that the popup could be closed before the completion of the commit creation allowing the user to see the program commit as normal.

Now, the popup is closed first so that awaiting commit creation does not effect the popup closing.