Made onCompareListScrolled callback optional
This callback is no longer required. If the there is not a need for this callback to store the new scroll position the callback does not have to be supplied.

Previously any use of the commit-list would need to implement the callback regardless of whether it gets use. So it has been made optional to avoid this.