Enter on a Blank-Slate cloneable repository list item now initiates clone
Previously if one selected a repository to clone and hit enter nothing would happen.

For a user it would be convenient to be able to initiate a clone by keyboard. Now, if a user hits enter on one of the repository list items the clone window will come up as it would if they clicked the clone button.