RetryCloneDialog no longer accepts optional messages/retryActions
The retry clone dialog has become more multi-purpose. Instead of replacing errors and showing it's own text, it now will display the error text directly. It will display any clone failure error and offer to retry cloning if applicable.

If there is no retry error, the retry dialog will not be used, and that error will instead be passed to the default handler.

Messages are no longer optional as well. However, if the message contains no text it will use a fallback text.