Fix accidental refactoring bug
In 81497244febb2a32b6b613e5ad9b72fa33aa5451 we were shuffling some code around and in the proccess we inadvertently moved the check for whether the main window exists to after we've destroyed it and cleaned up the handle.

The end result being that we always assume that a crash is the result of a launch problem.