Atomically delete remote ref
This changes the logic of deleteBranch from first checking if the remote ref exists and then depending on that operation deciding whether to attempt to delete the remote ref to instead always attempting to delete and then figuring out if the deletion failed because the remote ref didn't exist.

This will save us one (fairly expensive) network operation (ls-remote) while at the same time ensuring that the remote ref doesn't show up in the branch list when it couldn't be removed from the remote.