Use a standard Map instead of memoizeOne
memoizeOne() only caches the latest execution of that function, which means that when switching branches a few times we'll end up doing multiple requests. With the Map() approach we make sure to cache the information as long as needed (at the expenses of using more memory).

The `Map()` object will grow indefinitely, which could cause a theoretical memory leak. In practical terms I don't expect this object to hold a lot of information, since the amount of unpushed tags for any repository should be relatively small (and we only store a string for the tag name for each of them).