Replace forceFetch param by a clearCache() method
Using `forceFetch` when doing a new fetch caused an unintended issue: only the tags to push data for the current branch got refreshed, so if the user switched branches they would need to fetch again, which should be unnecessary.

Now, when the user pushes/pulls, we clear the cache with all tags to push for the specified remote.