Split out `findGitOnPath()` from `isGitOnPath()`
An upcoming patch will want to find out _where_ on the PATH the Git executable lives, not just that it is there. The `isGitOnPath()` function almost did that already, except that it just ignored the standard output of the command that found the path.

Note: there is a shortcut in `isGitOnPath()` for macOS, where it expects the executable to be on the PATH, and does not even look for it. We keep that shortcut there, but teach the `findGitOnPath()` function to handle macOS, too, just in case that this functionality will be needed any time in the future.