Remove obsolete user map
We've gradually over a number of commits worked our way to the userMap not being required any more.

It used to be used to lookup a GitHubUser from an email address but we're now relying entirely on the mentionables API and the improved logic of the Avatar component for the things we used to depend on.