Remove skip from test
It's better to test the head than to not test anything.

Unfortunately it's quite complex to fix the test. The problem comes from the `temp` NPM package, which uses the `TEMP` env variable in Windows machines.

For compatibility reasons, the `TEMP` env var in Windows returns the shortland DOS 8.3 version of the temp folder, which means that folders/files with more than 8 characters get a different name. In GitHub Actions the username is `runneradmin` which gets translated in DOS 8.3 to `RUNNER~1`.

The ideal solution would be to translate the shortland names back their full name be on the `setupEmptyRepository()` but it's quite tricky to do that (we would need to spawn `cmd` and do nasty internal windows calls).

Since this is the only test checking a path I think it's more pragmatic to just remove the path check.