Use absolute positioning of tooltips
Turns out fixed positioning doesn't always mean fixed in relation to the viewport, it means fixed in relation to the closest element which has `transform`, `perspective`, `filter` set to something other than `none` or has `will-change` set to `transform`.

See https://developer.mozilla.org/en-US/docs/Web/CSS/position#fixed

As it happens, we have a rule which sets `transform` to `none` on dialogs when the user prefers reduced motion which in turn made the tooltips positioned relative to the viewport instead of the dialog.

Instead of trying to fix the fixed positioning I'm opting here to go for absolute positioning instead. The tooltips will always be a direct child of the tooltip host so we don't have to worry about another element appearing in the ancestry creating a different frame of reference.