Refactor files list to use shared `onOpenInExternalEditor` fn
This refactor ensures that all double click event handlers bubble up to the same `onOpenInExternalEditor` fn in app.tsx.

This is where we prefix the repository path to the relative path which fixes a bug we caught where wrong files would open the opened project had a file at the same relative path (ie. README.md).