Make our tooltip a bit more aria-describedby friendly :joy:
Before this change, the tooltip would override (setting or removing) whatever the target element had set as `aria-describedby`.

This was evident when trying to set a custom `aria-describedby` to an element with a tooltip: the `aria-describedby` attribute wouldn't make it to the DOM :expressionless:

With this change, the tooltip will add or remove itself, respecting whatever was in there before.